Add constructor api to Aura
