+++ srml/aura/Cargo.toml
@@ -17,6 +17,7 @@ timestamp = { package = "srml-timestamp", path = "../timestamp", default-feature
 staking = { package = "srml-staking", path = "../staking", default-features = false }
 session = { package = "srml-session", path = "../session", default-features = false }
 substrate-consensus-aura-primitives = { path = "../../core/consensus/aura/primitives", default-features = false}
+safety-primitives = { package = "substrate-consensus-safety-primitives", path = "../../core/consensus/accountable-safety/primitives" }
 
 [dev-dependencies]
 lazy_static = "1.0"
